chore: 跳ね返るボールを表示
