+++ frontend/src/features/game/routes/Canvas.tsx
@@ -1,39 +1,13 @@
-import { memo, FC, useRef, useEffect, useCallback } from 'react';
+import { memo, FC } from 'react';
+import { useCanvas } from '../hooks/useCanvas';
 
-export const Canvas: FC = memo((props) => {
-  const canvasRef = useRef<HTMLCanvasElement>(null);
+type Props = {
+  draw: (ctx: CanvasRenderingContext2D, frameCount: number) => void;
+};
 
-  const draw = useCallback(
-    (ctx: CanvasRenderingContext2D, frameCount: number) => {
-      ctx.clearRect(0, 0, ctx.canvas.width, ctx.canvas.height);
-      ctx.fillStyle = '#000000';
-      ctx.beginPath();
-      ctx.arc(50, 100, 20 * Math.sin(frameCount * 0.05) ** 2, 0, 2 * Math.PI);
-      ctx.fill();
-    },
-    []
-  );
+export const Canvas: FC<Props> = memo((props) => {
+  const { draw } = props;
+  const canvasRef = useCanvas(draw);
 
-  useEffect(() => {
-    const canvas = canvasRef.current as HTMLCanvasElement;
-    const context = canvas.getContext('2d') as CanvasRenderingContext2D;
-
-    let frameCount = 0;
-    let animationFrameId = 0;
-
-    // Our draw came here
-    const render = () => {
-      frameCount++;
-      draw(context, frameCount);
-      animationFrameId = window.requestAnimationFrame(render);
-    };
-
-    render();
-
-    return () => {
-      window.cancelAnimationFrame(animationFrameId);
-    };
-  }, [draw]);
-
-  return <canvas ref={canvasRef} {...props} />;
+  return <canvas width={1000} height={500} ref={canvasRef} />;
 });
